pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fixdoc"
version = "0.0.3"
description = "Capture and search infrastructure fixes for cloud/SRE engineers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    { name="Fiyi", email="fiyignk@gmail.com" }
]
keywords = ["devops", "terraform", "infrastructure", "documentation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
fixdoc = "fixdoc.fix:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[project.urls]
Homepage = "https://github.com/fiyiogunkoya/FixDoc"
